requirements.txt
requests
python-docx
pandas
pdfplumber
openpyxl
